requirements.txt
robotframework==7.0
selenium==4.17.2
robotframework-seleniumlibrary==6.2.0